WIP: Introduce IntervalDigest type
 - IntervalDigest captures the result of a "namespaced hashing" op.
 - add proof verification API (WIP)
 - add Get method to get the leaves per namespace
 - fix some typos

+++ defaulthasher/hasher_test.go
@@ -23,7 +23,7 @@ func Test_namespacedTreeHasher_HashLeaf(t *testing.T) {
 
 	tests := []struct {
 		name  string
-		nsLen int
+		nsLen uint8
 		leaf  []byte
 		want  []byte
 	}{
@@ -52,7 +52,7 @@ func Test_namespacedTreeHasher_HashNode(t *testing.T) {
 
 	tests := []struct {
 		name     string
-		nidLen   int
+		nidLen   uint8
 		children children
 		want     []byte
 	}{
